refactor to shortnames, don't offer namespaced module names but provide backward compat for existing users, drop unused functions, don't duplicate headers, remove .keep files, fix version_added to align with Content Eng
