+++ tests/integration/targets/rule/tasks/main.yml
@@ -1,13 +1,13 @@
 ---
 - name: get info about qradar offense
-  qradar_rule_info:
+  rule_info:
     name: "Service DoS Attack Detected"
   register: rule_info
 
 - debug: var=rule_info
 
 - name: enable rule by id
-  qradar_rule:
+  rule:
     state: enabled
     id: "{{ rule_info['rules'][0]['id']|int }}"
   register: enable_rule_by_id
@@ -19,7 +19,7 @@
       - enable_rule_by_id is not failed
 
 - name: enable rule by id again
-  qradar_rule:
+  rule:
     state: enabled
     id: "{{ rule_info['rules'][0]['id']|int }}"
   register: enable_rule_by_id_again
@@ -31,7 +31,7 @@
       - enable_rule_by_id_again is not failed
 
 - name: disable rule by id
-  qradar_rule:
+  rule:
     state: disabled
     id: "{{ rule_info['rules'][0]['id']|int }}"
   register: disable_rule_by_id
@@ -43,7 +43,7 @@
       - disable_rule_by_id is not failed
 
 - name: disable rule by id again
-  qradar_rule:
+  rule:
     state: disabled
     id: "{{ rule_info['rules'][0]['id']|int }}"
   register: disable_rule_by_id_again
